Fix, but error while execute crawling.
```
objc[7785]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called.
objc[7785]: +[__NSCFConstantString initialize] may have been in progress in another thread when fork() was called. We cannot safely call it or ignore it in the fork() child process. Crashing instead. Set a breakpoint on objc_initializeAfterForkError to debug.
```
