+++ config/develop_config.py
@@ -1,0 +1,3 @@
+SCRAPY_SETTINGS_MODULE = 'src.service.crawler.crawler.settings'
+DEBUG = True
+DATABASE_URI =
