--- develop_config.py
@@ -1,1 +1,0 @@
-SCRAPY_SETTINGS_MODULE = 'service.crawler.crawler.settings'
