+++ requirements.txt
@@ -1,6 +1,48 @@
-Flask~=2.0.2
-Scrapy~=2.5.1
-scrapy-splash~=0.8.0
-SQLAlchemy~=1.4.29
-billiard~=3.6.4.0
-pymysql~=1.0.2
+aniso8601==9.0.1
+attrs==21.3.0
+Automat==20.2.0
+billiard==3.6.4.0
+cffi==1.15.0
+click==8.0.3
+constantly==15.1.0
+cryptography==36.0.1
+cssselect==1.1.0
+Flask==2.0.2
+greenlet==1.1.2
+h2==3.2.0
+hpack==3.0.0
+hyperframe==5.2.0
+hyperlink==21.0.0
+idna==3.3
+incremental==21.3.0
+itemadapter==0.4.0
+itemloaders==1.0.4
+itsdangerous==2.0.1
+Jinja2==3.0.3
+jmespath==0.10.0
+jsonschema==4.3.2
+lxml==4.7.1
+MarkupSafe==2.0.1
+parsel==1.6.0
+priority==1.3.0
+Protego==0.1.16
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser==2.21
+PyDispatcher==2.0.5
+PyMySQL==1.0.2
+pyOpenSSL==21.0.0
+pyrsistent==0.18.0
+python-dotenv==0.19.2
+pytz==2021.3
+queuelib==1.6.2
+Scrapy==2.5.1
+scrapy-splash==0.8.0
+service-identity==21.1.0
+six==1.16.0
+SQLAlchemy==1.4.29
+Twisted==21.7.0
+typing-extensions==4.0.1
+w3lib==1.22.0
+Werkzeug==2.0.2
+zope.interface==5.4.0
