+++ src/__init__.py
@@ -1,0 +1,3 @@
+from flask import Flask
+
+app = Flask(__name__)
