+++ src/service/crawling_service.py
@@ -1,8 +1,12 @@
+from typing import List
+
 from billiard.context import Process
 from billiard.dummy import Manager
 
+from src.service.board_data_service import get_active_codes
+from src.service.crawler.crawler import spiders
 from src.service.crawler.crawler.spiders import crawl_spider
-from src.service.crawler.custom_crawler import get_spider, CustomCrawler, get_scrapy_settings
+from src.service.crawler.custom_crawler import CustomCrawler
 
 
 def crawling_task(page_num, code):
@@ -14,7 +18,6 @@ def crawling_task(page_num, code):
     proc = Process(
         target=cc.crawling_start,
         args=(
-            get_scrapy_settings(),
             spider,
             code,
             return_dic,
@@ -28,9 +31,15 @@ def crawling_task(page_num, code):
             proc.join()
             is_success = True
         except Exception as e:
-            print(e)
-            print("WWWWWWWWWWWWWWWW")
             try_time += 1
     if try_time == 2:
         raise Exception(f"사이트에 연결하지 못했습니다. {spider.__name__}")
     return dict(return_dic)
+
+
+def get_spider(code) -> object:
+    return eval(f"spiders.{code.capitalize()}Spider")
+
+
+def get_all_spider() -> List[object]:
+    return [eval(f"spiders.{code.capitalize()}Spider") for code in get_active_codes()]
